Rename: `reserve`또는 `index` 페이지에서 사용하는 컴포넌트들을 구분하기 위한 폴더 분기
